cleaned up code, formatted config cells and view
