--- Braze-Demo/Model/Content-Cards/GroupSection.swift
@@ -1,7 +1,0 @@
-enum GroupSection: Int, CaseIterable {
-  case blank
-  case primary
-  case secondary
-  case headline
-  case large
-}
