+++ Braze-Demo/Storyboard/Main.storyboard
@@ -570,6 +570,50 @@
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="5AR-RX-77L">
                                 <rect key="frame" x="0.0" y="88" width="414" height="725"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                                <prototypes>
+                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="none" indentationWidth="10" reuseIdentifier="ConfigTableViewCell" rowHeight="102" id="Tri-xJ-Zn8" customClass="ConfigTableViewCell" customModule="Braze_Demo" customModuleProvider="target">
+                                        <rect key="frame" x="0.0" y="44.5" width="414" height="102"/>
+                                        <autoresizingMask key="autoresizingMask"/>
+                                        <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="Tri-xJ-Zn8" id="SkK-NE-OIM">
+                                            <rect key="frame" x="0.0" y="0.0" width="414" height="102"/>
+                                            <autoresizingMask key="autoresizingMask"/>
+                                            <subviews>
+                                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Lorem" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="5Pb-Q2-GrQ">
+                                                    <rect key="frame" x="15" y="10" width="49.5" height="20.5"/>
+                                                    <fontDescription key="fontDescription" type="system" weight="medium" pointSize="17"/>
+                                                    <nil key="textColor"/>
+                                                    <nil key="highlightedColor"/>
+                                                </label>
+                                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Lorem" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="sQw-Pv-PNM">
+                                                    <rect key="frame" x="15" y="35.5" width="40.5" height="17"/>
+                                                    <fontDescription key="fontDescription" type="system" pointSize="14"/>
+                                                    <nil key="textColor"/>
+                                                    <nil key="highlightedColor"/>
+                                                </label>
+                                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Lorem" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ufm-Ty-xzr">
+                                                    <rect key="frame" x="15" y="76" width="38" height="16"/>
+                                                    <fontDescription key="fontDescription" type="system" pointSize="13"/>
+                                                    <nil key="textColor"/>
+                                                    <nil key="highlightedColor"/>
+                                                </label>
+                                            </subviews>
+                                            <constraints>
+                                                <constraint firstItem="ufm-Ty-xzr" firstAttribute="leading" secondItem="sQw-Pv-PNM" secondAttribute="leading" id="2vw-8w-wsu"/>
+                                                <constraint firstAttribute="bottom" secondItem="ufm-Ty-xzr" secondAttribute="bottom" constant="10" id="CEy-IK-Yo6"/>
+                                                <constraint firstItem="sQw-Pv-PNM" firstAttribute="leading" secondItem="5Pb-Q2-GrQ" secondAttribute="leading" id="Npm-rj-cO3"/>
+                                                <constraint firstItem="5Pb-Q2-GrQ" firstAttribute="leading" secondItem="SkK-NE-OIM" secondAttribute="leading" constant="15" id="Ud9-5T-kcN"/>
+                                                <constraint firstItem="5Pb-Q2-GrQ" firstAttribute="top" secondItem="SkK-NE-OIM" secondAttribute="top" constant="10" id="UiO-Ek-DYH"/>
+                                                <constraint firstItem="sQw-Pv-PNM" firstAttribute="top" secondItem="5Pb-Q2-GrQ" secondAttribute="bottom" constant="5" id="Zl5-mw-ahh"/>
+                                                <constraint firstItem="ufm-Ty-xzr" firstAttribute="top" relation="greaterThanOrEqual" secondItem="sQw-Pv-PNM" secondAttribute="bottom" constant="20" id="o3M-F7-a9f"/>
+                                            </constraints>
+                                        </tableViewCellContentView>
+                                        <connections>
+                                            <outlet property="identifierLabel" destination="sQw-Pv-PNM" id="8Ml-wP-owq"/>
+                                            <outlet property="titleLabel" destination="5Pb-Q2-GrQ" id="Moq-5g-siH"/>
+                                            <outlet property="updatedLabel" destination="ufm-Ty-xzr" id="64A-dS-jIF"/>
+                                        </connections>
+                                    </tableViewCell>
+                                </prototypes>
                                 <connections>
                                     <outlet property="delegate" destination="EDM-4f-Xv3" id="16R-KM-XyV"/>
                                 </connections>
@@ -591,7 +635,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Pzc-Ze-70i" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="4500" y="812"/>
+            <point key="canvasLocation" x="4498.5507246376819" y="811.60714285714278"/>
         </scene>
         <!--Content Cards-->
         <scene sceneID="y9X-Hy-8vG">
